requirements.txt
altair==4.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
streamlit==1.41.1
captcha==0.7.1